requirements.txt
Django>=3.2,<4.0
gunicorn>=20.1
psycopg2-binary
python-json-logger>=2.0.0,<3.0.0
djangorestframework
django-filter
python-dotenv
python-decouple